SCB-1670 refactor pack-tests name because it's a IT for RestTemplate scenario
