+++ integration-tests/pack-tests/pom.xml
@@ -27,7 +27,7 @@
   <modelVersion>4.0.0</modelVersion>
 
   <artifactId>pack-tests</artifactId>
-  <name>Pack::Integration Tests</name>
+  <name>Pack::Integration Tests::RestTemplate</name>
 
 
   <dependencyManagement>
