+++ integration-tests/pack-tests/src/test/java/org/apache/servicecomb/pack/integration/tests/resttemplate/TxEventEnvelopeRepository.java
@@ -6,7 +6,7 @@
  * (the "License"); you may not use this file except in compliance with
  * the License.  You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ *      http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -15,10 +15,17 @@
  * limitations under the License.
  */
 
-package org.apache.servicecomb.pack.integration.tests;
+package org.apache.servicecomb.pack.integration.tests.resttemplate;
 
-import org.apache.servicecomb.pack.alpha.core.Command;
+import java.util.List;
+
+import org.apache.servicecomb.pack.alpha.core.TxEvent;
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 
-interface CommandEnvelopeRepository extends CrudRepository<Command, Long> {
+interface TxEventEnvelopeRepository extends CrudRepository<TxEvent, Long> {
+  List<TxEvent> findByGlobalTxIdOrderByCreationTime(String globalTxId);
+
+  @Query("SELECT DISTINCT(e.globalTxId) from TxEvent e order by e.creationTime desc")
+  List<String> findDistinctGlobalTxId();
 }
